[trainer] refactor: simplify custom pipeline adapters
Consolidate the QwenImage training and rollout adapters under installable verl_omni custom pipeline modules so scripts, tests, and docs reference canonical paths. This removes duplicated scheduler and example implementations while keeping the custom pipeline layout easier to extend to new model families.

Made-with: Cursor
