+++ examples/flowgrpo_trainer/run_qwen_image_ocr_lora.sh
@@ -26,7 +26,7 @@ python3 -m verl_omni.trainer.main_flowgrpo \
     data.max_prompt_length=256 \
     actor_rollout_ref.model.path=${MODEL_PATH:-$WORKSPACE/models/Qwen/Qwen-Image} \
     actor_rollout_ref.model.tokenizer_path=${ACTOR_TOKENIZER_PATH:-$WORKSPACE/models/Qwen/Qwen-Image/tokenizer} \
-    actor_rollout_ref.model.external_lib="verl_omni.custom_pipelines.qwen_image.diffusers_impl" \
+    actor_rollout_ref.model.external_lib="verl_omni.custom_pipelines.qwen_image.diffusers_training_adapter" \
     actor_rollout_ref.model.lora_rank=64 \
     actor_rollout_ref.model.lora_alpha=128 \
     actor_rollout_ref.model.target_modules="['to_q','to_k','to_v','to_out.0','add_q_proj','add_k_proj','add_v_proj','to_add_out','img_mlp.net.0.proj','img_mlp.net.2','txt_mlp.net.0.proj','txt_mlp.net.2']" \
@@ -53,7 +53,7 @@ python3 -m verl_omni.trainer.main_flowgrpo \
     actor_rollout_ref.rollout.algo.sde_window_range="[0,5]" \
     actor_rollout_ref.rollout.val_kwargs.num_inference_steps=50 \
     actor_rollout_ref.rollout.val_kwargs.algo.noise_level=0.0 \
-    actor_rollout_ref.rollout.external_lib=examples.flowgrpo_trainer.vllm_omni_impl \
+    actor_rollout_ref.rollout.external_lib=verl_omni.custom_pipelines.qwen_image.vllm_omni_rollout_adapter \
     actor_rollout_ref.ref.log_prob_micro_batch_size_per_gpu=32 \
     reward.num_workers=$((NUM_GPUS_ACTOR_ROLLOUT_REWARD / REWARD_TP)) \
     reward.reward_manager.name=visual \
