+++ examples/flowgrpo_trainer/vllm_omni_impl/__init__.py
@@ -11,6 +11,6 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from .pipeline_qwenimage import QwenImagePipelineWithLogProb
+from verl_omni.custom_pipelines.qwen_image.vllm_omni_rollout_adapter import QwenImagePipelineWithLogProb
 
 __all__ = ["QwenImagePipelineWithLogProb"]
