+++ tests/experimental/agent_loop/test_diffusion_agent_loop.py
@@ -86,7 +86,7 @@ def init_config() -> DictConfig:
         config.actor_rollout_ref.rollout.max_sequence_length = max_length
         config.actor_rollout_ref.rollout.nnodes = 1
 
-        config.actor_rollout_ref.rollout.external_lib = "examples.flowgrpo_trainer.vllm_omni_impl"
+        config.actor_rollout_ref.rollout.external_lib = "verl_omni.custom_pipelines.qwen_image.vllm_omni_rollout_adapter"
         config.reward.reward_manager.name = "image"
         config.trainer.n_gpus_per_node = 4
